search: Added an option to show highlight of match
There are 2 new options -g and -G. -g accepts a parameter whereas -G does not.
-G would only show the highlight of the text found in the file content if it
exists, whereas -g allows to specify a value to highlight in the highlighted
text. This would typically be the same as the query, but since queries can be
complicated that is not necessary true.
Also removed extra new lines in the README.md file.